augmented_dataset
In this particular commit, the dataset has been finally prepared for training the model. Augmentation of the dataset is performed to create a plagiarised version of the data. Further , this commit contains changes in feature extraction and a separate file to check the number of files in each dataset for debugging and understanding.